requirements.txt
acme==1.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anthropic==0.59.0
anyio==4.9.0
appdirs==1.4.4
apsw==3.36.0.post1
arrow==1.2.1
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7
autogenstudio==0.4.2.2
Automat==20.2.0
azure-ai-agents==1.0.2
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b12
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.23.1
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
Babel==2.8.0
backcall==0.2.0
backoff==2.2.1
baostock==0.8.9
bcrypt==4.3.0
beautifulsoup4==4.13.4
beniget==0.4.1
blinker==1.9.0
Brotli==1.0.9
browser-use==0.1.40
browserbase==1.4.0
build==1.2.2.post1
cachetools==5.5.2
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2025.7.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.2.1
cloud-init==24.4
cobble==0.1.4
cohere==5.16.1
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
crewai==0.150.0
crewai-chat-ui==0.1.2
crewai-tools==0.58.0
cryptography==45.0.5
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
cycler==0.11.0
dataclasses-json==0.6.7
dbus-python==1.2.18
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
email_validator==2.2.0
embedchain==0.1.128
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastavro==1.11.1
feedparser==6.0.8
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.29.1
frozenlist==1.7.0
fs==2.4.12
fsspec==2025.3.0
gast==0.5.2
google-auth==2.40.3
googleapis-common-protos==1.70.0
gptcache==0.1.44
gradio==5.21.0
gradio_client==1.7.2
greenlet==3.1.1
groovy==0.1.2
grpcio==1.74.0
h11==0.14.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
ifaddr==0.1.7
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==21.3.0
instructor==1.10.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.9.0
josepy==1.10.0
json5==0.12.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==2.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==23.5.0
kiwisolver==1.3.2
kubernetes==33.1.0
lancedb==0.24.2
langchain==0.3.27
langchain-anthropic==0.3.3
langchain-cohere==0.3.5
langchain-community==0.3.27
langchain-core==0.3.72
langchain-deepseek==0.1.2
langchain-experimental==0.3.4
langchain-ollama==0.2.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.9
langsmith==0.3.45
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.8.1
litellm==1.74.3
loguru==0.7.3
lxml==4.8.0
lz4==3.1.3+dfsg
magika==0.6.2
Mako==1.3.10
mammoth==1.9.1
Markdown==3.3.6
markdown-it-py==3.0.0
markdownify==0.14.1
markitdown==0.1.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.5.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mechanize==0.4.7
mem0ai==0.1.115
mmh3==5.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.0.3
multidict==6.6.3
mypy_extensions==1.1.0
mysqlclient==1.4.6
netaddr==0.8.0
netifaces==0.11.0
networkx==3.5
nodeenv==1.9.1
numpy==2.2.4
oauthlib==3.3.1
olefile==0.46
ollama==0.4.7
onnxruntime==1.22.0
openai==1.97.1
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
orjson==3.11.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
playwright==1.50.0
ply==3.11
portalocker==2.7.0
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psycopg==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==0.11.3+dfsg
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pychm==0.8.6
pycparser==2.22
pycryptodomex==3.11.0
pycurl==7.44.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.10.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==5.9.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyright==1.1.403
pyrsistent==0.18.1
pysbd==0.3.4
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pythran==0.10.0
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
pyxdg==0.27
PyYAML==6.0.2
qdrant-client==1.15.0
referencing==0.36.2
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.9
rignore==0.6.4
Routes==2.5.1
rpds-py==0.26.0
rsa==4.9.1
ruff==0.11.0
safehttpx==0.1.6
schema==0.7.7
scipy==1.8.0
SecretStorage==3.3.1
semantic-version==2.10.0
sentry-sdk==2.33.2
service-identity==18.1.0
setuptools==76.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlmodel==0.0.24
ssh-import-id==5.11
stack-data==0.6.3
stagehand==0.4.1
starlette==0.46.1
sympy==1.14.0
systemd-python==234
tabulate==0.9.0
tenacity==9.1.2
texttable==1.6.4
tiktoken==0.9.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==22.1.0
typer==0.15.2
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.5.0
uv==0.8.3
uvicorn==0.34.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.37.1
wrapt==1.17.2
xdg==5
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
youtube-transcript-api==1.0.3
zeroconf==0.38.3
zipp==3.23.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
zstandard==0.23.0
